<docs>(widgets): 위젯 개선
