+++ lib/ui/pages/ex_register/ex_register_page.dart
@@ -4,7 +4,7 @@ import 'package:artlog_app_mvp/ui/pages/now/now_checkin_page.dart';
 import 'package:artlog_app_mvp/ui/widgets/common/image_uploader.dart';
 import 'package:flutter/material.dart';
 import 'package:artlog_app_mvp/ui/widgets/appbars/custom_appbar.dart';
-import 'package:artlog_app_mvp/ui/widgets/cards/%08common_card.dart';
+import 'package:artlog_app_mvp/ui/widgets/cards/input_card.dart';
 import 'package:artlog_app_mvp/ui/widgets/buttons/contained_button.dart';
 import 'package:intl/intl.dart';
