+++ lib/ui/pages/now/now_checkin_page.dart
@@ -1,7 +1,7 @@
 import 'package:artlog_app_mvp/ui/pages/ex_register/ex_date_%08selection_sheet.dart';
 import 'package:artlog_app_mvp/ui/pages/now/now_review_page.dart';
 import 'package:artlog_app_mvp/ui/widgets/buttons/contained_button.dart';
-import 'package:artlog_app_mvp/ui/widgets/cards/%08common_card.dart';
+import 'package:artlog_app_mvp/ui/widgets/cards/input_card.dart';
 import 'package:artlog_app_mvp/ui/widgets/common/image_uploader.dart';
 import 'package:artlog_app_mvp/ui/widgets/icons/icon_widgets.dart';
 import 'package:flutter/material.dart';
