+++ lib/ui/widgets/common/record_badge.dart
@@ -13,20 +13,19 @@ class RecordBadge extends StatelessWidget {
     final Map<RecordBadgeType, Map<String, dynamic>> styles = {
       RecordBadgeType.deep: {
         "label": "DEEP",
-        "backgroundColor": Color(0x190770E8),
-        "textColor": Color(0xFF0770E8),
-        "width": 36.0,
+        "backgroundColor": Color(0xE8E9F7FF),  
+        "textColor": Color(0xFF0341D1), 
       },
       RecordBadgeType.look: {
         "label": "LOOK",
-        "backgroundColor": Color(0x1900E068),
-        "textColor": Color(0xFF019D4A),
+        "backgroundColor": Color(0xD1D3F0F1),  
+        "textColor": Color(0xFF00C967), 
         "width": 38.0,
       },
       RecordBadgeType.now: {
         "label": "NOW",
-        "backgroundColor": Color(0x19FF459E),
-        "textColor": Color(0xFFFF459E),
+        "backgroundColor": Color(0xFFFFF0F4),
+        "textColor": Color(0xFFFF52A5),  
         "width": 35.0,
       },
     };
@@ -51,6 +50,7 @@ class RecordBadge extends StatelessWidget {
               fontSize: 11,
               fontFamily: 'Pretendard',
               fontWeight: FontWeight.w700,
+              letterSpacing: 0,  // No letter-spacing
             ),
           ),
         ],
